Bug 1293798 – Move credential storage into PushManager
